+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ0AH0RHSQ5BZ29AWHCSQND6"
+        "01KJ0C15HRG66KY2T4S80TZW7C"
     )
     r"""The unique identifier of the data source"""