+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -186,7 +186,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kj0ah0875z2n8sph3zvkt3xq"
+        "conv_01kj0c1511rmw23w0bxxezpc1g"
     )
     r"""Unique conversation identifier with `conv_` prefix."""