+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveidentityop.py
@@ -103,7 +103,7 @@ class RetrieveIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-21T14:43:58.401Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-21T15:10:16.389Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")