+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ0AH0RM32PBCDMEEWJ0RMEC"
+        "01KJ0C15HWB1RGGJGKJKJMNA5E"
     )
     r"""The unique identifier of the data source"""