+++ packages/orq-rc/RELEASES.md
@@ -5929,3 +5929,13 @@ Based on:
 - [python v4.4.0-rc.25] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.25 - packages/orq-rc
+
+## 2026-02-21 15:13:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.723.0 (2.835.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.26] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.26 - packages/orq-rc